Revisions to pull request
Dictionaries for SSPk and ARS schemes are now stored with their classes

Explicit parts of SSPk schemes are now available as a tableau via a new factory.
These tableau were also added to the testing in test_explicit.py

Testing in test_imex.py is now parameterized by the Butcher tableau
